refactor: criando feature de proposal
