+++ frontend/src/features/freelancers/pages/proposal/CadastrarProposta.tsx
@@ -30,8 +30,8 @@ import { useCreateProposal } from "../../hooks/useCreateProposal";
 import type { CriarPropostaDTO } from "../../dtos/freelancer/PropostaDTO";
 import { parseCurrencyToNumber } from "@/shared/utils/number.utils";
 import { useNavigate, useParams } from "react-router-dom";
-import type { ServicoDTO } from "@/features/services/dtos/ServicoDTO";
-import { consultarServicoPorId } from "@/features/services/api/servico.api";
+import type { ServicoDTO } from "@/features/services/proposal/dtos/ServicoDTO";
+import { consultarServicoPorId } from "@/features/services/proposal/api/servico.api";
 import Spinner from "@/shared/components/ui/Spinner";
 import type { ClienteDTO } from "@/features/clients/dtos/ClienteDTO";
 import { consultarClientePorId } from "@/features/clients/api/cliente.api";
