+++ frontend/src/features/services/proposal/routes/proposalPaths.ts
@@ -1,0 +1,7 @@
+export const proposalPaths = {
+  pesquisaServico: "/pesquisa-servico",
+  visualizarServico: "/visualizar-servico",
+  cadastrarServico: "/cadastrar-servico",
+
+  visualizarServicoById: (id: string | number) => `/visualizar-servico/${id}`,
+};
