+++ frontend/src/features/services/routes/servico.routes.tsx
@@ -1,29 +1,15 @@
 import { RequireRole } from "@/features/auth/components/RequireRole";
-import { CadastrarServico } from "@/features/services/pages/CadastrarServico";
-import PesquisaServico from "@/features/services/pages/PesquisaServicos";
-import { VisualizarServico } from "@/features/services/pages/VisualizarServico";
+import { CadastrarServico } from "@/features/services/proposal/pages/CadastrarServico";
+import PesquisaServico from "@/features/services/proposal/pages/PesquisaServicos";
+import { VisualizarServico } from "@/features/services/proposal/pages/VisualizarServico";
 import { MeuServico } from "@/features/services/pages/MeuServico";
 import { servicoPaths } from "@/features/services/routes/servicoPaths";
 
 export const servicoRoutes = [
-  { path: servicoPaths.pesquisaServico, element: <PesquisaServico /> },
-
-  {
-    path: `${servicoPaths.visualizarServico}/:id`,
-    element: <VisualizarServico />,
-  },
 
   {
     path: `${servicoPaths.meuServico}/:id`,
     element: <MeuServico />,
   },
 
-  {
-    path: servicoPaths.cadastrarServico,
-    element: (
-      <RequireRole roles={["Cliente"]}>
-        <CadastrarServico />
-      </RequireRole>
-    ),
-  },
 ];
